Move ToSpecifiedFullPath() to a shared file. Expand the comments
