+++ src/System.IO.FileSystem/src/System/IO/Enumeration/FileSystemEntry.Unix.cs
@@ -9,7 +9,7 @@ namespace System.IO.Enumeration
     /// <summary>
     /// Lower level view of FileSystemInfo used for processing and filtering find results.
     /// </summary>
-    public unsafe ref struct FileSystemEntry
+    public unsafe ref partial struct FileSystemEntry
     {
         private const int FileNameBufferSize = 256;
         internal Interop.Sys.DirectoryEntry _directoryEntry;
@@ -135,25 +135,6 @@ namespace System.IO.Enumeration
         {
             string fullPath = ToFullPath();
             return FileSystemInfo.Create(fullPath, new string(FileName), ref _status);
-        }
-
-        /// <summary>
-        /// Returns the full path for find results, based on the initially provided path.
-        /// </summary>
-        public string ToSpecifiedFullPath()
-        {
-            // We want to provide the enumerated segment of the path appended to the originally specified path. This is
-            // the behavior of the various Directory APIs that return a list of strings.
-            //
-            // RootDirectory has the final separator trimmed, OriginalRootDirectory does not. Our legacy behavior would
-            // effectively account for this by appending subdirectory names as it recursed. As such we need to trim one
-            // separator when combining with the relative path (Directory.Slice(RootDirectory.Length)).
-
-            ReadOnlySpan<char> relativePath = Directory.Slice(RootDirectory.Length);
-            if (PathInternal.EndsInDirectorySeparator(OriginalRootDirectory) && PathInternal.StartsWithDirectorySeparator(relativePath))
-                relativePath = relativePath.Slice(1);
-
-            return Path.Join(OriginalRootDirectory, relativePath, FileName);
         }
 
         /// <summary>
